Remove unreachable generic FROST protocol scaffolding
