+++ pkg/frost/signing/native_frost_dkg_engine_frost_native.go
@@ -2,8 +2,6 @@
 
 package signing
 
-import "fmt"
-
 // NativeFROSTDKGRound1Package is the public package broadcast during FROST DKG
 // round one.
 type NativeFROSTDKGRound1Package struct {
@@ -18,8 +16,8 @@ type NativeFROSTDKGRound2Package struct {
 	// native DKG package.
 	Identifier string `json:"identifier"`
 	// SenderIdentifier is filled by the Go coordinator for packages received
-	// from peers. UniFFI Part3 needs to key round-two packages by the sender
-	// while the package itself carries the recipient.
+	// from peers. The tbtc-signer DKG Part3 request keys round-two packages by
+	// sender while the package itself carries the recipient.
 	SenderIdentifier string `json:"senderIdentifier,omitempty"`
 	Data             []byte `json:"data"`
 }
@@ -53,77 +51,4 @@ type NativeFROSTDKGPart2Result struct {
 type NativeFROSTDKGResult struct {
 	KeyPackage       *NativeFROSTKeyPackage       `json:"keyPackage"`
 	PublicKeyPackage *NativeFROSTPublicKeyPackage `json:"publicKeyPackage"`
-}
-
-// SignerMaterial rejects the unsupported generic UniFFI FROST DKG output.
-// FROST wallet material must be persisted through the tbtc-signer engine so
-// Taproot tweaked signing is available for deposit sweeps.
-func (nfdkg *NativeFROSTDKGResult) SignerMaterial() (*NativeSignerMaterial, error) {
-	if nfdkg == nil {
-		return nil, fmt.Errorf("native FROST DKG result is nil")
-	}
-
-	return nil, fmt.Errorf(
-		"native FROST DKG result cannot be persisted as unsupported [%s] signer material; use [%s]",
-		NativeSignerMaterialFormatFrostUniFFIV2,
-		NativeSignerMaterialFormatFrostTBTCSignerV1,
-	)
-}
-
-// NativeFROSTDKGEngine executes the cryptographic primitives for the three
-// FROST DKG parts. It intentionally exposes only serializable package data to
-// the coordinator; the bridge implementation is responsible for adapting these
-// values to the underlying UniFFI/tbtc-signer handle model.
-type NativeFROSTDKGEngine interface {
-	Part1(
-		participantIdentifier string,
-		maxSigners uint16,
-		minSigners uint16,
-	) (*NativeFROSTDKGPart1Result, error)
-	Part2(
-		secretPackage *NativeFROSTDKGRound1SecretPackage,
-		round1Packages []*NativeFROSTDKGRound1Package,
-	) (*NativeFROSTDKGPart2Result, error)
-	Part3(
-		secretPackage *NativeFROSTDKGRound2SecretPackage,
-		round1Packages []*NativeFROSTDKGRound1Package,
-		round2Packages []*NativeFROSTDKGRound2Package,
-	) (*NativeFROSTDKGResult, error)
-}
-
-var nativeFROSTDKGEngine NativeFROSTDKGEngine
-
-// RegisterNativeFROSTDKGEngine registers the native FROST DKG cryptographic
-// engine used by the FROST wallet-registry coordinator.
-func RegisterNativeFROSTDKGEngine(engine NativeFROSTDKGEngine) error {
-	if engine == nil {
-		return fmt.Errorf("native FROST DKG engine is nil")
-	}
-
-	executionBackendMutex.Lock()
-	defer executionBackendMutex.Unlock()
-
-	nativeFROSTDKGEngine = engine
-
-	return nil
-}
-
-// UnregisterNativeFROSTDKGEngine clears native FROST DKG engine registration.
-func UnregisterNativeFROSTDKGEngine() {
-	executionBackendMutex.Lock()
-	defer executionBackendMutex.Unlock()
-
-	nativeFROSTDKGEngine = nil
-}
-
-func currentNativeFROSTDKGEngine() NativeFROSTDKGEngine {
-	executionBackendMutex.RLock()
-	defer executionBackendMutex.RUnlock()
-
-	return nativeFROSTDKGEngine
-}
-
-// CurrentNativeFROSTDKGEngine returns the registered native FROST DKG engine.
-func CurrentNativeFROSTDKGEngine() NativeFROSTDKGEngine {
-	return currentNativeFROSTDKGEngine()
 }
